feat: cart and order resources and daos.

+++ Coursework/bookstore/src/main/java/dev/isuru/resource/AuthorResource.java
@@ -1,10 +1,13 @@
 package dev.isuru.resource;
 
 import dev.isuru.dao.AuthorDAO;
+import dev.isuru.exception.AuthorNotFoundException;
 import dev.isuru.model.Author;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import javax.validation.Valid;
+import javax.validation.constraints.NotNull;
 import javax.ws.rs.*;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
@@ -26,6 +29,9 @@ public class AuthorResource {
     @Produces(MediaType.APPLICATION_JSON)
     @Path("/{id}")
     public Response getAuthorById(@PathParam("id") int id) {
+        if (!authorDAO.contains(id)) {
+            throw new AuthorNotFoundException(id);
+        }
         Response response = Response.ok(authorDAO.get(id)).build();
         logger.info("{} GET authors/{}", response.getStatus(), id);
         return response;
@@ -34,9 +40,9 @@ public class AuthorResource {
     @POST
     @Produces(MediaType.APPLICATION_JSON)
     @Consumes(MediaType.APPLICATION_JSON)
-    public Response createAuthor(Author author) {
+    public Response createAuthor(@NotNull @Valid Author author) {
         authorDAO.add(author);
-        Response response = Response.status(Response.Status.CREATED).build();
+        Response response = Response.status(Response.Status.CREATED).entity(author).build();
         logger.info("{} POST authors/", response.getStatus());
         return response;
     }
@@ -45,7 +51,10 @@ public class AuthorResource {
     @Produces(MediaType.APPLICATION_JSON)
     @Consumes(MediaType.APPLICATION_JSON)
     @Path("/{id}")
-    public Response updateAuthor(@PathParam("id") int id, Author author) {
+    public Response updateAuthor(@PathParam("id") int id, @NotNull @Valid Author author) {
+        if (!authorDAO.contains(id)) {
+            throw new AuthorNotFoundException(id);
+        }
         authorDAO.update(id, author);
         Response response = Response.status(Response.Status.NO_CONTENT).build();
         logger.info("{} PUT authors/{}", response.getStatus(), id);
@@ -56,6 +65,9 @@ public class AuthorResource {
     @Produces(MediaType.APPLICATION_JSON)
     @Path("/{id}")
     public Response deleteAuthor(@PathParam("id") int id) {
+        if (!authorDAO.contains(id)) {
+            throw new AuthorNotFoundException(id);
+        }
         authorDAO.delete(id);
         Response response = Response.status(Response.Status.NO_CONTENT).build();
         logger.info("{} DELETE authors/{}", response.getStatus(), id);
